MATCHER/requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
apscheduler
surprise
pandas
numpy